requirements.txt
python-dotenv
rank_bm25
langchain
langchain-community
langchain-huggingface
langchain-groq
pypdf
faiss-cpu==1.10.0